- client: Initial swipe at automatic proxy server configuration         detection on Windows.  Fixes #35
    client/
        cs_cmdline.cpp
        http_curl.cpp
        http_curl.h
    lib/
        boinc_win.h
        proxy_info.cpp
        proxy_info.h
        util.cpp
        win_util.h

svn path=/branches/boinc_core_release_6_8/; revision=18773

+++ lib/util.cpp
@@ -323,7 +323,7 @@ int read_file_malloc(const char* path, char*& buf, size_t max_len, bool tail) {
 #ifndef _USING_FCGI_
     if (max_len && size > max_len) {
         if (tail) {
-            fseek(f, (long)size-max_len, SEEK_SET);
+            fseek(f, (long)size-(long)max_len, SEEK_SET);
         }
         size = max_len;
     }
